+++ src/components/action-popover/index.js
@@ -1,3 +1,4 @@
 export { default as ActionPopover } from './action-popover.component';
+export { default as ActionPopoverMenu } from './action-popover-menu.component';
 export { default as ActionPopoverItem } from './action-popover-item.component';
 export { default as ActionPopoverDivider } from './action-popover-divider.component';
